+++ backend/utils/bool.go
@@ -1,0 +1,9 @@
+package utils
+
+func BoolToInt(val bool) int {
+	conv := 0
+	if val {
+		conv = 1
+	}
+	return conv
+}